feat(top-nav) : 네비게이션 바에 기능 추가
- 로그인과 뒤로가기 버튼을 추가하고 기능 구현

- 프로필 아이콘 클릭 시, 로그인 화면으로 / 뒤로가기 아이콘 클릭시, 이전 화면으로 이동

- 폴더 위치 변경이 이루어짐 (공통으로 사용하는 컴포넌트 폴더에 위치)
